+++ docs/spec/src/index.ts
@@ -1,9 +1,16 @@
 import { IntegrationSpecConfig } from '@jupiterone/integration-sdk-core';
 
 import { IntegrationConfig } from '../../../src/config';
-import { accessSpec } from './access';
 import { accountSpec } from './account';
+import { applicationSpec } from './application';
+import { deviceSpec } from './device';
+import { userSpec } from './user';
 
 export const invocationConfig: IntegrationSpecConfig<IntegrationConfig> = {
-  integrationSteps: [...accountSpec, ...accessSpec],
+  integrationSteps: [
+    ...accountSpec,
+    ...applicationSpec,
+    ...deviceSpec,
+    ...userSpec,
+  ],
 };