+++ src/steps/constants.ts
@@ -6,74 +6,56 @@ import {
 
 export const Steps = {
   ACCOUNT: 'fetch-account',
+  APPLICATIONS: 'fetch-applications',
+  DEVICES: 'fetch-devices',
   USERS: 'fetch-users',
-  GROUPS: 'fetch-groups',
-  GROUP_USER_RELATIONSHIPS: 'build-user-group-relationships',
 };
 
 export const Entities: Record<
-  'ACCOUNT' | 'GROUP' | 'USER',
+  'ACCOUNT' | 'APPLICATION' | 'DEVICE' | 'USER',
   StepEntityMetadata
 > = {
   ACCOUNT: {
     resourceName: 'Account',
-    _type: 'acme_account',
+    _type: 'simplemdm_account',
     _class: ['Account'],
-    schema: {
-      properties: {
-        mfaEnabled: { type: 'boolean' },
-        manager: { type: 'string' },
-      },
-      required: ['mfaEnabled', 'manager'],
-    },
   },
-  GROUP: {
-    resourceName: 'UserGroup',
-    _type: 'acme_group',
-    _class: ['UserGroup'],
-    schema: {
-      properties: {
-        email: { type: 'string' },
-        logoLink: { type: 'string' },
-      },
-      required: ['email', 'logoLink'],
-    },
+  APPLICATION: {
+    resourceName: 'Application',
+    _type: 'simplemdm_application',
+    _class: ['Application'],
+  },
+  DEVICE: {
+    resourceName: 'Device',
+    _type: 'simplemdm_device',
+    _class: ['Device'],
   },
   USER: {
     resourceName: 'User',
-    _type: 'acme_user',
+    _type: 'simplemdm_user',
     _class: ['User'],
-    schema: {
-      properties: {
-        username: { type: 'string' },
-        email: { type: 'string' },
-        active: { type: 'boolean' },
-        firstName: { type: 'string' },
-      },
-      required: ['username', 'email', 'active', 'firstName'],
-    },
   },
 };
 
 export const Relationships: Record<
-  'ACCOUNT_HAS_USER' | 'ACCOUNT_HAS_GROUP' | 'GROUP_HAS_USER',
+  'ACCOUNT_HAS_APPLICATION' | 'ACCOUNT_HAS_DEVICE' | 'DEVICE_HAS_USER',
   StepRelationshipMetadata
 > = {
-  ACCOUNT_HAS_USER: {
-    _type: 'acme_account_has_user',
+  ACCOUNT_HAS_APPLICATION: {
+    _type: 'simplemdm_account_has_application',
     sourceType: Entities.ACCOUNT._type,
     _class: RelationshipClass.HAS,
-    targetType: Entities.USER._type,
+    targetType: Entities.APPLICATION._type,
   },
-  ACCOUNT_HAS_GROUP: {
-    _type: 'acme_account_has_group',
+  ACCOUNT_HAS_DEVICE: {
+    _type: 'simplemdm_account_has_device',
     sourceType: Entities.ACCOUNT._type,
     _class: RelationshipClass.HAS,
-    targetType: Entities.GROUP._type,
+    targetType: Entities.DEVICE._type,
   },
-  GROUP_HAS_USER: {
-    _type: 'acme_group_has_user',
-    sourceType: Entities.GROUP._type,
+  DEVICE_HAS_USER: {
+    _type: 'simplemdm_device_has_user',
+    sourceType: Entities.DEVICE._type,
     _class: RelationshipClass.HAS,
     targetType: Entities.USER._type,
   },