+++ src/steps/index.ts
@@ -1,6 +1,13 @@
 import { accountSteps } from './account';
-import { accessSteps } from './access';
+import { applicationSteps } from './application';
+import { deviceSteps } from './device';
+import { userSteps } from './user';
 
-const integrationSteps = [...accountSteps, ...accessSteps];
+const integrationSteps = [
+  ...accountSteps,
+  ...applicationSteps,
+  ...deviceSteps,
+  ...userSteps,
+];
 
 export { integrationSteps };