+++ CHANGELOG.md
@@ -7,3 +7,15 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+
+- Ingest new entities
+  - `simplemdm_account`
+  - `simplemdm_application`
+  - `simplemdm_device`
+  - `simplemdm_user`
+- Build new relationships
+  - `simplemdm_account_has_application`
+  - `simplemdm_account_has_device`
+  - `simplemdm_device_has_user`